+++ backend/src/main/java/reviewme/reviewgroup/domain/exception/ReviewGroupNotFoundByReviewException.java
@@ -1,4 +1,4 @@
-package reviewme.review.service.exception;
+package reviewme.reviewgroup.domain.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import reviewme.global.exception.DataInconsistencyException;
