+++ backend/src/main/java/reviewme/reviewgroup/domain/exception/ReviewGroupUnAuthorizedException.java
@@ -1,7 +1,6 @@
-package reviewme.review.service.exception;
+package reviewme.reviewgroup.domain.exception;
 
 import lombok.extern.slf4j.Slf4j;
-import reviewme.global.exception.BadRequestException;
 import reviewme.global.exception.UnAuthorizedException;
 
 @Slf4j
